refactor: rename Box component to BlockDisplay
