+++ packages/app/src/components/canvas/BlockDisplay.tsx
